sample-apps/summarization-gcs/requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-aiplatform